SE API: hide private interfaces
- Split each headers into module.h and module_priv.h, move the methods that
  is only used internally by SE implementation to module_priv.h, and export
  module_priv.h to rest of TEE Core
- Added new include path to se_api_self_tests.c for which needs to include
  private headers
- Split aid.c and apdu.c from iso7816.c. Originally they have to be wriiten in
  the same file since they share some private data structures. Now, the
  private data structure can be shared via private headers.
- Split reader.c from manager.c for the same reason above.

Tested-by: SY Chiu <sy.chiu@linaro.org> (Modified QEMU + jcardsim)
