+++ core/arch/arm32/sta/sub.mk
@@ -3,4 +3,5 @@ srcs-y += core_self_tests.c
 
 ifeq ($(WITH_SE_API),y)
 srcs-${WITH_SE_API_SELF_TEST} += se_api_self_tests.c
+cppflags-se_api_self_tests.c-y += -Icore/tee/se
 endif
