+++ core/include/tee/se/manager.h
@@ -36,6 +36,9 @@ TEE_Result tee_se_manager_get_readers(
 		struct tee_se_reader_proxy **proxy_list,
 		size_t *proxy_list_size);
 
+bool tee_se_manager_is_reader_proxy_valid(
+		struct tee_se_reader_proxy *proxy);
+
 size_t tee_se_manager_get_reader_count(void);
 
 #endif
