+++ core/include/tee/se/reader.h
@@ -32,29 +32,14 @@
 #include <kernel/mutex.h>
 #include <sys/queue.h>
 
+struct tee_se_reader_proxy;
 struct tee_se_session;
-
-/*
- * Reader Proxy is used to serialize access from multiple seesions,
- * and maintain reference counter. All access to the reader should
- * go through Reader Proxy
- */
-struct tee_se_reader_proxy {
-	struct tee_se_reader *reader;
-	int refcnt;
-	bool basic_channel_locked;
-	struct mutex mutex;
-
-	TAILQ_ENTRY(tee_se_reader_proxy) link;
-};
 
 TEE_Result tee_se_reader_get_name(struct tee_se_reader_proxy *proxy,
 		char **reader_name, size_t *reader_name_len);
 
 void tee_se_reader_get_properties(struct tee_se_reader_proxy *proxy,
 		TEE_SEReaderProperties *prop);
-
-int tee_se_reader_get_refcnt(struct tee_se_reader_proxy *proxy);
 
 TEE_Result tee_se_reader_attach(struct tee_se_reader_proxy *proxy);
 
@@ -71,14 +56,10 @@ TEE_Result tee_se_reader_get_atr(struct tee_se_reader_proxy *proxy,
 TEE_Result tee_se_reader_transmit(struct tee_se_reader_proxy *proxy,
 		uint8_t *tx_buf, size_t tx_buf_len, uint8_t *rx_buf, size_t *rx_buf_len);
 
-TEE_Result tee_se_reader_check_state(struct tee_se_reader_proxy *proxy);
-
 void tee_se_reader_lock_basic_channel(struct tee_se_reader_proxy *proxy);
 
 void tee_se_reader_unlock_basic_channel(struct tee_se_reader_proxy *proxy);
 
 bool tee_se_reader_is_basic_channel_locked(struct tee_se_reader_proxy *proxy);
-
-bool tee_se_reader_is_proxy_valid(struct tee_se_reader_proxy *proxy);
 
 #endif
