+++ core/tee/se/channel_priv.h
@@ -25,48 +25,23 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef TEE_SE_SERVICE_H
-#define TEE_SE_SERVICE_H
+#ifndef TEE_SE_CHANNEL_PRIV_H
+#define TEE_SE_CHANNEL_PRIV_H
 
-#include <tee_api_types.h>
-#include <kernel/mutex.h>
+struct tee_se_aid;
 
-struct tee_se_service;
-struct tee_se_session;
-struct tee_se_channel;
-struct tee_se_reader_proxy;
+struct tee_se_channel {
+	int channel_id;
+	struct tee_se_session *session;
+	struct tee_se_aid *aid;
+	struct resp_apdu *select_resp;
 
-TEE_Result tee_se_service_open(
-		struct tee_se_service **service);
+	TAILQ_ENTRY(tee_se_channel) link;
+};
 
-TEE_Result tee_se_service_add_session(
-		struct tee_se_service *service,
-		struct tee_se_session *session);
+struct tee_se_channel *tee_se_channel_alloc(struct tee_se_session *s,
+		int channel_id);
 
-void tee_se_service_close_session(
-		struct tee_se_service *service,
-		struct tee_se_session *session);
-
-void tee_se_service_close_sessions_by_reader(
-		struct tee_se_service *service,
-		struct tee_se_reader_proxy *proxy);
-
-TEE_Result tee_se_service_is_session_closed(
-		struct tee_se_service *service,
-		struct tee_se_session *session_service);
-
-TEE_Result tee_se_service_close(
-		struct tee_se_service *service);
-
-bool tee_se_service_is_valid(
-		struct tee_se_service *service);
-
-bool tee_se_service_is_session_valid(
-		struct tee_se_service *service,
-		struct tee_se_session *session_service);
-
-bool tee_se_service_is_channel_valid(
-		struct tee_se_service *service,
-		struct tee_se_channel *channel);
+void tee_se_channel_free(struct tee_se_channel *c);
 
 #endif
