+++ core/tee/se/reader_priv.h
@@ -25,48 +25,25 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef TEE_SE_SERVICE_H
-#define TEE_SE_SERVICE_H
+#ifndef TEE_SE_READER_PRIV_H
+#define TEE_SE_READER_PRIV_H
 
-#include <tee_api_types.h>
-#include <kernel/mutex.h>
+/*
+ * Reader Proxy is used to serialize access from multiple seesions,
+ * and maintain reference counter. All access to the reader should
+ * go through Reader Proxy
+ */
+struct tee_se_reader_proxy {
+	struct tee_se_reader *reader;
+	int refcnt;
+	bool basic_channel_locked;
+	struct mutex mutex;
 
-struct tee_se_service;
-struct tee_se_session;
-struct tee_se_channel;
-struct tee_se_reader_proxy;
+	TAILQ_ENTRY(tee_se_reader_proxy) link;
+};
 
-TEE_Result tee_se_service_open(
-		struct tee_se_service **service);
+TEE_Result tee_se_reader_check_state(struct tee_se_reader_proxy *proxy);
 
-TEE_Result tee_se_service_add_session(
-		struct tee_se_service *service,
-		struct tee_se_session *session);
-
-void tee_se_service_close_session(
-		struct tee_se_service *service,
-		struct tee_se_session *session);
-
-void tee_se_service_close_sessions_by_reader(
-		struct tee_se_service *service,
-		struct tee_se_reader_proxy *proxy);
-
-TEE_Result tee_se_service_is_session_closed(
-		struct tee_se_service *service,
-		struct tee_se_session *session_service);
-
-TEE_Result tee_se_service_close(
-		struct tee_se_service *service);
-
-bool tee_se_service_is_valid(
-		struct tee_se_service *service);
-
-bool tee_se_service_is_session_valid(
-		struct tee_se_service *service,
-		struct tee_se_session *session_service);
-
-bool tee_se_service_is_channel_valid(
-		struct tee_se_service *service,
-		struct tee_se_channel *channel);
+int tee_se_reader_get_refcnt(struct tee_se_reader_proxy *proxy);
 
 #endif
