+++ core/tee/se/service_priv.h
@@ -25,48 +25,18 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef TEE_SE_SERVICE_H
-#define TEE_SE_SERVICE_H
+#ifndef TEE_SE_SERVICE_PRIV_H
+#define TEE_SE_SERVICE_PRIV_H
 
-#include <tee_api_types.h>
-#include <kernel/mutex.h>
+TAILQ_HEAD(se_session_head, tee_se_session);
 
-struct tee_se_service;
-struct tee_se_session;
-struct tee_se_channel;
-struct tee_se_reader_proxy;
-
-TEE_Result tee_se_service_open(
-		struct tee_se_service **service);
-
-TEE_Result tee_se_service_add_session(
-		struct tee_se_service *service,
-		struct tee_se_session *session);
-
-void tee_se_service_close_session(
-		struct tee_se_service *service,
-		struct tee_se_session *session);
-
-void tee_se_service_close_sessions_by_reader(
-		struct tee_se_service *service,
-		struct tee_se_reader_proxy *proxy);
-
-TEE_Result tee_se_service_is_session_closed(
-		struct tee_se_service *service,
-		struct tee_se_session *session_service);
-
-TEE_Result tee_se_service_close(
-		struct tee_se_service *service);
-
-bool tee_se_service_is_valid(
-		struct tee_se_service *service);
-
-bool tee_se_service_is_session_valid(
-		struct tee_se_service *service,
-		struct tee_se_session *session_service);
-
-bool tee_se_service_is_channel_valid(
-		struct tee_se_service *service,
-		struct tee_se_channel *channel);
+struct tee_se_service {
+	/* list of sessions opened on the service */
+	struct se_session_head opened_sessions;
+	/* list of sessions closed on the service */
+	struct se_session_head closed_sessions;
+	/* mutex to pretect the session lists */
+	struct mutex mutex;
+};
 
 #endif
