+++ core/tee/se/session_priv.h
@@ -25,48 +25,22 @@
  * POSSIBILITY OF SUCH DAMAGE.
  */
 
-#ifndef TEE_SE_SERVICE_H
-#define TEE_SE_SERVICE_H
+#ifndef TEE_SE_SESSION_PRIV_H
+#define TEE_SE_SESSION_PRIV_H
 
-#include <tee_api_types.h>
-#include <kernel/mutex.h>
+TAILQ_HEAD(channel_list, tee_se_channel);
 
-struct tee_se_service;
-struct tee_se_session;
-struct tee_se_channel;
-struct tee_se_reader_proxy;
+struct tee_se_session {
+	struct tee_se_reader_proxy *reader_proxy;
 
-TEE_Result tee_se_service_open(
-		struct tee_se_service **service);
+	/* list of channels opened on the session*/
+	struct channel_list channels;
 
-TEE_Result tee_se_service_add_session(
-		struct tee_se_service *service,
-		struct tee_se_session *session);
+	TAILQ_ENTRY(tee_se_session) link;
+};
 
-void tee_se_service_close_session(
-		struct tee_se_service *service,
-		struct tee_se_session *session);
+struct tee_se_session *tee_se_session_alloc(struct tee_se_reader_proxy *proxy);
 
-void tee_se_service_close_sessions_by_reader(
-		struct tee_se_service *service,
-		struct tee_se_reader_proxy *proxy);
-
-TEE_Result tee_se_service_is_session_closed(
-		struct tee_se_service *service,
-		struct tee_se_session *session_service);
-
-TEE_Result tee_se_service_close(
-		struct tee_se_service *service);
-
-bool tee_se_service_is_valid(
-		struct tee_se_service *service);
-
-bool tee_se_service_is_session_valid(
-		struct tee_se_service *service,
-		struct tee_se_session *session_service);
-
-bool tee_se_service_is_channel_valid(
-		struct tee_se_service *service,
-		struct tee_se_channel *channel);
+void tee_se_session_free(struct tee_se_session *s);
 
 #endif
