+++ core/tee/se/sub.mk
@@ -1,3 +1,3 @@
-srcs-y += service.c manager.c iso7816.c session.c channel.c util.c svc.c
+srcs-y += service.c manager.c reader.c iso7816.c session.c channel.c aid.c apdu.c util.c svc.c
 
 subdirs-y += reader
